fix interger to float

+++ db/migrate/20151005194206_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration
     create_table :products do |t|
       t.string :account_name
       t.integer :total_fees
-      t.integer :healthscore
+      t.float :healthscore
 
       t.timestamps null: false
     end
